EducAI/backend/requirements.txt
Flask==2.1.1
openai==1.34.0
azure-core==1.23.1
azure-search-documents==11.2.0
gunicorn==20.1.0
Werkzeug==2.2.2
